moved the postcall for api to the thoughtpage-component and added styling for the displayed messages
