+++ src/assets/components/PostThought.css
@@ -1,18 +1,26 @@
 .post-container {
-    background: lightgray;
-    margin: 0 auto;
+    background: #ededed;
     width: 500px;
     height: 200px;
     border: 1px solid black;
+    box-shadow: 7px 7px;
+    margin-top: 30px;
 }
 
 .post-container button {
-    padding: 10px;
+    width: 210px;
+    padding: 12px;
     margin: 10px;
+    margin-left: 20px;
     border: none;
     border-radius: 20px;
-    background-color: plum;
-    color: red;
+    background-color: #ff9696;
+    font-size: 15px;
+}
+
+.post-container button:hover {
+    box-shadow: 0 0 13px rgba(33, 33, 33, .2);
+    transition: 0.3s;
 }
 
 .question-container {
@@ -21,14 +29,25 @@
 }
 
 .question-container h1 {
-    font-size: 20px;
+    margin-left: 20px;
+    font-size: 17px;
 }
 
 .input-container {
     text-align: center;
+
 }
 
 .input-container input {
-    width: 80%;
-    height: 80px;
+    width: 90%;
+    height: 60px;
+    font-size: 16px;
+    font-family: "Inconsolata", monospace;
+    font-weight: 600;
+}
+
+input::placeholder {
+    font-size: 16px;
+    font-family: "Inconsolata", monospace;
+    font-weight: 500;
 }
